+++ clients/ivs.d.ts
@@ -515,6 +515,10 @@ declare namespace IVS {
      */
     recordingReconnectWindowSeconds?: RecordingReconnectWindowSeconds;
     /**
+     * Object that describes which renditions should be recorded for a stream.
+     */
+    renditionConfiguration?: RenditionConfiguration;
+    /**
      * Array of 1-50 maps, each of the form string:string (key:value). See Tagging Amazon Web Services Resources for more information, including restrictions that apply to tags and "Tag naming limits and requirements"; Amazon IVS has no service-specific constraints beyond what is documented there.
      */
     tags?: Tags;
@@ -911,6 +915,10 @@ declare namespace IVS {
      */
     recordingReconnectWindowSeconds?: RecordingReconnectWindowSeconds;
     /**
+     * Object that describes which renditions should be recorded for a stream.
+     */
+    renditionConfiguration?: RenditionConfiguration;
+    /**
      * Indicates the current state of the recording configuration. When the state is ACTIVE, the configuration is ready for recording a channel stream.
      */
     state: RecordingConfigurationState;
@@ -951,6 +959,19 @@ declare namespace IVS {
   }
   export type RecordingMode = "DISABLED"|"INTERVAL"|string;
   export type RecordingReconnectWindowSeconds = number;
+  export interface RenditionConfiguration {
+    /**
+     * Indicates which set of renditions are recorded for a stream. For BASIC channels, the CUSTOM value has no effect. If CUSTOM is specified, a set of renditions must be specified in the renditions field. Default: ALL.
+     */
+    renditionSelection?: RenditionConfigurationRenditionSelection;
+    /**
+     * Indicates which renditions are recorded for a stream, if renditionSelection is CUSTOM; otherwise, this field is irrelevant. The selected renditions are recorded if they are available during the stream. If a selected rendition is unavailable, the best available rendition is recorded. For details on the resolution dimensions of each rendition, see Auto-Record to Amazon S3.
+     */
+    renditions?: RenditionConfigurationRenditionList;
+  }
+  export type RenditionConfigurationRendition = "FULL_HD"|"HD"|"SD"|"LOWEST_RESOLUTION"|string;
+  export type RenditionConfigurationRenditionList = RenditionConfigurationRendition[];
+  export type RenditionConfigurationRenditionSelection = "ALL"|"NONE"|"CUSTOM"|string;
   export type ResourceArn = string;
   export type S3DestinationBucketName = string;
   export interface S3DestinationConfiguration {
@@ -1177,10 +1198,21 @@ declare namespace IVS {
      */
     recordingMode?: RecordingMode;
     /**
-     * The targeted thumbnail-generation interval in seconds. This is configurable (and required) only if recordingMode is INTERVAL. Default: 60.  Important: Setting a value for targetIntervalSeconds does not guarantee that thumbnails are generated at the specified interval. For thumbnails to be generated at the targetIntervalSeconds interval, the IDR/Keyframe value for the input video must be less than the targetIntervalSeconds value. See  Amazon IVS Streaming Configuration for information on setting IDR/Keyframe to the recommended value in video-encoder settings.
+     * Indicates the desired resolution of recorded thumbnails. Thumbnails are recorded at the selected resolution if the corresponding rendition is available during the stream; otherwise, they are recorded at source resolution. For more information about resolution values and their corresponding height and width dimensions, see Auto-Record to Amazon S3. Default: Null (source resolution is returned).
+     */
+    resolution?: ThumbnailConfigurationResolution;
+    /**
+     * Indicates the format in which thumbnails are recorded. SEQUENTIAL records all generated thumbnails in a serial manner, to the media/thumbnails directory. LATEST saves the latest thumbnail in media/latest_thumbnail/thumb.jpg and overwrites it at the interval specified by targetIntervalSeconds. You can enable both SEQUENTIAL and LATEST. Default: SEQUENTIAL.
+     */
+    storage?: ThumbnailConfigurationStorageList;
+    /**
+     * The targeted thumbnail-generation interval in seconds. This is configurable (and required) only if recordingMode is INTERVAL. Default: 60.  Important: For the BASIC channel type, setting a value for targetIntervalSeconds does not guarantee that thumbnails are generated at the specified interval. For thumbnails to be generated at the targetIntervalSeconds interval, the IDR/Keyframe value for the input video must be less than the targetIntervalSeconds value. See  Amazon IVS Streaming Configuration for information on setting IDR/Keyframe to the recommended value in video-encoder settings.
      */
     targetIntervalSeconds?: TargetIntervalSeconds;
   }
+  export type ThumbnailConfigurationResolution = "FULL_HD"|"HD"|"SD"|"LOWEST_RESOLUTION"|string;
+  export type ThumbnailConfigurationStorage = "SEQUENTIAL"|"LATEST"|string;
+  export type ThumbnailConfigurationStorageList = ThumbnailConfigurationStorage[];
   export type Time = Date;
   export type TranscodePreset = "HIGHER_BANDWIDTH_DELIVERY"|"CONSTRAINED_BANDWIDTH_DELIVERY"|string;
   export interface UntagResourceRequest {